Revert "Fix more Java warnings."
This reverts commit a290f1c1f9ead7e7173ab3e4546a5c9fbe1d57d2.

+++ src/main/java/edu/umn/ecology/populus/model/aspg/ASPGPanel.java
@@ -587,7 +587,7 @@ public class ASPGPanel extends BasicPlotInputPanel implements java.io.Externaliz
 
     void tableModel_tableChanged(TableModelEvent e) {
         if (e.getColumn() == 1 && e.getLastRow() == 0 && getModelType() == kPREBREEDING) {
-            this.initialL1PPF.setCurrentValue((Double) tableModel.getValueAt(0, 1));
+            this.initialL1PPF.setCurrentValue(((Double) tableModel.getValueAt(0, 1)).doubleValue());
         }
     }
